refactor(quil-node): complete Batch 3 (subsystem init extraction)
Extracts the remaining run_master_node init phases per
.claude/plans/deep-bouncing-cerf.md. No behavior changes.

- master_node/runtime_state.rs: Phases 5c+5d (message collector +
  prover registry + coverage monitor + halt state + halt broadcaster +
  prover-only ticker). Returns RuntimeState. ~220 lines extracted.
- master_node/worker_manager.rs: Phase 5e (worker manager — cluster
  vs local thread mode, drain task, persisted-state restore,
  pre-allocation, config filters, pi_worker_manager OnceLock publish).
  ~600 lines extracted. drain task stays inline per pragmatic scoping
  (no separate worker_drain.rs).
- master_node/allocator_and_lifecycle.rs: Phases 5g/h/i (worker
  allocator + seniority compat + early reconcile spawn + consensus
  OnceLocks + prover lifecycle + frame materializer). ~226 lines
  extracted.

main.rs: 4,862 → 3,874 lines. cumulative reduction from original
6,208 is 2,334 lines (37.6%).

Verification: cargo check --workspace, cargo nextest run --workspace
(2,326 passed / 13 skipped), simtest/ci.sh -short — all green.
