+++ crates/quil-node/src/main.rs
@@ -443,1055 +443,67 @@ async fn run_master_node(
         archive_mode,
     });
 
-    // ---------------------------------------------------------------
-    // 5c. Message collector + consensus event loop
-    // ---------------------------------------------------------------
-    let message_collector = Arc::new(quil_engine::message_collector::MessageCollector::new());
+    let master_node::runtime_state::RuntimeState {
+        message_collector,
+        prover_registry,
+        prover_only_flag: _prover_only_flag,
+        global_event_distributor,
+        coverage_monitor,
+        halt_state,
+        remote_worker_manager_for_halt,
+    } = master_node::runtime_state::init(hg_store.clone(), shard_engines.clone(), token.clone());
 
-    // The consensus event loop is optional — it runs only when this node
-    // has a BLS proving key and is an active prover. For now we create
-    // the prover registry and message collector, but defer the full
-    // HotStuff event loop until the node has synced enough state to
-    // determine its role. The consensus components are ready to be
-    // wired when this node becomes an active prover.
-    let prover_registry = Arc::new(quil_execution::SharedProverRegistry::new());
-    {
-        let pr = prover_registry.clone();
-        let hs = hg_store.clone();
-        // Run in background — don't block P2P startup. The registry
-        // populates asynchronously; consensus won't start until it's ready.
-        tokio::task::spawn_blocking(move || {
-            pr.refresh_from_store(&hs);
-            let count = pr.read(|r| r.distinct_provers());
-            tracing::info!(provers = count, "prover registry loaded (background)");
-        });
-    }
+    let worker_manager: Arc<dyn quil_engine::worker::WorkerManager> = master_node::worker_manager::init(
+        master_node::worker_manager::WorkerManagerArgs {
+            config: config.clone(),
+            archive_mode,
+            token: token.clone(),
+            p2p_handle: p2p_handle.clone(),
+            db_arc: db_arc.clone(),
+            clock_store: clock_store.clone(),
+            crdt: crdt.clone(),
+            exec_manager: exec_manager.clone(),
+            inclusion_prover: inclusion_prover.clone(),
+            frame_prover: frame_prover.clone(),
+            message_collector: message_collector.clone(),
+            fee_manager: fee_manager.clone(),
+            prover_registry: prover_registry.clone(),
+            halt_state: halt_state.clone(),
+            file_key_manager: file_key_manager.clone(),
+            prover_address,
+            bls_pubkey: bls_pubkey.clone(),
+            shard_engines: shard_engines.clone(),
+            remote_worker_manager_for_halt: remote_worker_manager_for_halt.clone(),
+            pi_worker_manager: pi_worker_manager.clone(),
+        },
+    );
 
-    // ---------------------------------------------------------------
-    // 5d. Coverage monitor + worker allocator + worker threads
-    // ---------------------------------------------------------------
-    let prover_only_flag = Arc::new(std::sync::atomic::AtomicBool::new(false));
-    let global_event_distributor: Arc<dyn quil_types::consensus::EventDistributor> =
-        Arc::new(quil_engine::event_distributor::InMemoryEventDistributor::new());
-    let coverage_monitor = Arc::new(quil_engine::coverage::CoverageMonitor::new(
-        prover_registry.clone() as Arc<dyn quil_types::consensus::ProverRegistry>,
-        global_event_distributor.clone(),
-        quil_engine::coverage::CoverageThresholds::mainnet(),
-        prover_only_flag.clone(),
-    ));
-
-    // Shared halt state — written by the coverage-event subscriber
-    // below, read by the lifecycle `join_proposal_ready` gate and by
-    // the periodic eviction scheduler. Needs to exist before any
-    // consumer references it.
-    let halt_state = Arc::new(quil_engine::halt_state::HaltState::new());
-
-    // Spawn subscriber: drains ControlEvents from the in-memory
-    // distributor and updates `halt_state` on CoverageHalt /
-    // CoverageResume. The separate shard orchestration subscriber is
-    // wired below once the prover pipeline exists.
-    {
-        let mut rx = global_event_distributor.subscribe("halt-state");
-        let hs = halt_state.clone();
-        let cancel = token.clone();
-        tokio::spawn(async move {
-            loop {
-                tokio::select! {
-                    biased;
-                    _ = cancel.cancelled() => break,
-                    maybe_event = rx.recv() => {
-                        let Some(event) = maybe_event else { break };
-                        match (event.event_type, &event.data) {
-                            (
-                                quil_types::consensus::ControlEventType::CoverageHalt,
-                                quil_types::consensus::ControlEventData::Coverage { filter, duration },
-                            ) => {
-                                if hs.apply(&event) {
-                                    info!(
-                                        filter = hex::encode(filter),
-                                        duration_frames = *duration,
-                                        halted_count = hs.halted_count(),
-                                        "coverage halt entered"
-                                    );
-                                    quil_engine::metrics::inc_coverage_halts_entered();
-                                    quil_engine::metrics::set_halted_shards(hs.halted_count() as u64);
-                                }
-                            }
-                            (
-                                quil_types::consensus::ControlEventType::Halt,
-                                quil_types::consensus::ControlEventData::Alert { message },
-                            ) => {
-                                // Hard halt — ALL shards stop. This is
-                                // a fire alarm, not a per-shard
-                                // coverage issue. Permanent for this
-                                // process lifetime; operator must
-                                // restart after the alert is resolved.
-                                error!(
-                                    message = %message,
-                                    "GLOBAL ALERT — hard halt, stopping all shard engines"
-                                );
-                                hs.hard_halt();
-                            }
-                            (
-                                quil_types::consensus::ControlEventType::CoverageResume,
-                                quil_types::consensus::ControlEventData::Coverage { filter, .. },
-                            ) => {
-                                if hs.apply(&event) {
-                                    info!(
-                                        filter = hex::encode(filter),
-                                        halted_count = hs.halted_count(),
-                                        "coverage halt resumed"
-                                    );
-                                    quil_engine::metrics::inc_coverage_resumes();
-                                    quil_engine::metrics::set_halted_shards(hs.halted_count() as u64);
-                                }
-                            }
-                            (
-                                quil_types::consensus::ControlEventType::ShardSplitEligible,
-                                quil_types::consensus::ControlEventData::ShardSplit { filter, proposed },
-                            ) => {
-                                info!(
-                                    filter = hex::encode(filter),
-                                    proposed = proposed.len(),
-                                    "shard split eligible (orchestration pending)"
-                                );
-                            }
-                            (
-                                quil_types::consensus::ControlEventType::ShardMergeEligible,
-                                quil_types::consensus::ControlEventData::ShardMerge { filters, parent },
-                            ) => {
-                                info!(
-                                    filter_count = filters.len(),
-                                    parent = hex::encode(parent),
-                                    "shard merge eligible (orchestration pending)"
-                                );
-                            }
-                            (quil_types::consensus::ControlEventType::CoverageWarn, _) => {
-                                debug!("coverage warn");
-                                quil_engine::metrics::inc_coverage_warns();
-                            }
-                            _ => {}
-                        }
-                    }
-                }
-            }
-        });
-    }
-
-    // Broadcast halt-state changes to every active app shard engine
-    // AND every standalone worker process. In-process thread-mode
-    // engines receive via the `shard_engines` map; standalone workers
-    // receive via the `SetHalted` DataIPC RPC, looked up through
-    // `remote_worker_manager_for_halt` (populated later by the
-    // cluster-mode branch of the worker_manager setup).
-    //
-    // Two firings:
-    //   1. Edge-triggered: when `halt_state.watch_any_halted()` flips,
-    //      push the new value immediately.
-    //   2. Periodic resync (every 30s) while a halt is active: covers
-    //      worker reconnects and the initial-connect race where a
-    //      worker boots mid-halt and would otherwise stay halted=false
-    //      until the next edge transition.
-    let remote_worker_manager_for_halt: Arc<
-        std::sync::OnceLock<Arc<quil_engine::remote_worker::RemoteWorkerManager>>,
-    > = Arc::new(std::sync::OnceLock::new());
-    {
-        let mut rx = halt_state.watch_any_halted();
-        let engines = shard_engines.clone();
-        let cancel = token.clone();
-        let remote_mgr_cell = remote_worker_manager_for_halt.clone();
-        let halt_state_for_periodic = halt_state.clone();
-        tokio::spawn(async move {
-            let mut periodic = tokio::time::interval(std::time::Duration::from_secs(30));
-            periodic.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
-            loop {
-                tokio::select! {
-                    _ = cancel.cancelled() => break,
-                    res = rx.changed() => {
-                        if res.is_err() { break; }
-                        let halted = *rx.borrow();
-                        // Scope the read guard: parking_lot::RwLockReadGuard
-                        // is !Send, so it must drop before any .await.
-                        let count = {
-                            let map = engines.read();
-                            let n = map.len();
-                            for (_filter, handle) in map.iter() {
-                                handle.set_halted(halted);
-                            }
-                            n
-                        };
-                        if let Some(mgr) = remote_mgr_cell.get() {
-                            mgr.broadcast_set_halted(halted).await;
-                        }
-                        info!(
-                            halted,
-                            engines = count,
-                            "broadcast halt state to app shard engines"
-                        );
-                    }
-                    _ = periodic.tick() => {
-                        let halted = halt_state_for_periodic.any_halted();
-                        if !halted {
-                            continue;
-                        }
-                        // Re-push to every engine + remote worker so a
-                        // freshly-connected worker picks up the halt.
-                        {
-                            let map = engines.read();
-                            for (_filter, handle) in map.iter() {
-                                handle.set_halted(true);
-                            }
-                        }
-                        if let Some(mgr) = remote_mgr_cell.get() {
-                            mgr.broadcast_set_halted(true).await;
-                        }
-                    }
-                }
-            }
-        });
-    }
-
-    // Wire prover-only mode into the message collector
-    // (the collector checks this flag on each add_message call)
-    let mc_prover_only = message_collector.clone();
-    let pof = prover_only_flag.clone();
-    tokio::spawn(async move {
-        let mut interval = tokio::time::interval(std::time::Duration::from_secs(1));
-        loop {
-            interval.tick().await;
-            mc_prover_only.set_prover_only_mode(
-                pof.load(std::sync::atomic::Ordering::Relaxed),
-            );
-        }
+    let master_node::allocator_and_lifecycle::LifecycleHandles {
+        worker_allocator,
+        consensus_handle,
+        vote_aggregator,
+        timeout_aggregator,
+        prover_lifecycle,
+        frame_materializer,
+    } = master_node::allocator_and_lifecycle::init(master_node::allocator_and_lifecycle::LifecycleInitArgs {
+        config: config.clone(),
+        network,
+        archive_mode,
+        token: token.clone(),
+        worker_manager: worker_manager.clone(),
+        prover_registry: prover_registry.clone(),
+        prover_address,
+        halt_state: halt_state.clone(),
+        current_frame: current_frame.clone(),
+        last_global_head_frame: last_global_head_frame.clone(),
+        shards_store: shards_store.clone(),
+        exec_manager: exec_manager.clone(),
+        frame_prover: frame_prover.clone(),
+        inclusion_prover: inclusion_prover.clone(),
+        clock_store: clock_store.clone(),
+        crdt: crdt.clone(),
+        hg_store: hg_store.clone(),
     });
-
-    // Worker manager — either local threads or remote gRPC workers.
-    // If data_worker_stream_multiaddrs has entries, use remote mode
-    // (cluster of machines). Otherwise, use local threads.
-    let reward_greedy = config.engine.reward_strategy == "reward-greedy";
-    let fkm_for_factory = file_key_manager.clone();
-
-    let worker_manager: Arc<dyn quil_engine::worker::WorkerManager> =
-        if !config.engine.data_worker_stream_multiaddrs.is_empty() {
-            // CLUSTER MODE: remote workers via gRPC
-            // Master listens on the stream port from P2P config
-            let master_port = if config.p2p.stream_listen_multiaddr.is_empty() {
-                8340u16
-            } else {
-                // Extract port from /ip4/X/tcp/PORT
-                config.p2p.stream_listen_multiaddr
-                    .split('/')
-                    .collect::<Vec<_>>()
-                    .windows(2)
-                    .find(|w| w[0] == "tcp")
-                    .and_then(|w| w[1].parse::<u16>().ok())
-                    .unwrap_or(8340)
-            };
-            let master_ep = format!("http://0.0.0.0:{}", master_port);
-            let remote_mgr = Arc::new(quil_engine::remote_worker::RemoteWorkerManager::from_config(
-                &config.engine.data_worker_stream_multiaddrs,
-                master_ep,
-            ));
-            info!(
-                remote_workers = config.engine.data_worker_stream_multiaddrs.len(),
-                "remote worker manager ready (cluster mode)"
-            );
-            // Publish to the halt broadcaster spawned above so it can
-            // SetHalted across standalone workers when coverage halts.
-            let _ = remote_worker_manager_for_halt.set(remote_mgr.clone());
-            remote_mgr as Arc<dyn quil_engine::worker::WorkerManager>
-        } else {
-            // LOCAL MODE: core-pinned threads
-            let thread_mgr = Arc::new(quil_engine::thread_worker::ThreadWorkerManager::new());
-            // Persistent worker registry — survives restarts so the
-            // operator's `manually_managed` flag and the
-            // worker→filter binding don't reset every reboot.
-            let worker_store: Arc<dyn quil_types::store::WorkerStore> =
-                Arc::new(quil_store::RocksWorkerStore::new(db_arc.inner()));
-            thread_mgr.set_worker_store(worker_store);
-            // Closure invoked by AppFollower from inside the consensus
-            // event loop: wraps a finalized FrameHeader (canonical
-            // bytes) in a `MessageBundle{Shard: header}` and publishes
-            // on `GLOBAL_PROVER`. Spawning the actual publish keeps the
-            // call non-blocking from the consensus side.
-            let coverage_p2p = p2p_handle.clone();
-            let coverage_publish: Arc<dyn Fn(Vec<u8>) + Send + Sync> =
-                Arc::new(move |header_canonical_bytes: Vec<u8>| {
-                    let req = match quil_execution::message_envelope::CanonicalMessageRequest::wrap(
-                        header_canonical_bytes,
-                    ) {
-                        Ok(r) => r,
-                        Err(e) => {
-                            warn!(error = %e, "coverage publish: bad FrameHeader bytes");
-                            return;
-                        }
-                    };
-                    let timestamp = std::time::SystemTime::now()
-                        .duration_since(std::time::UNIX_EPOCH)
-                        .unwrap_or_default()
-                        .as_millis() as i64;
-                    let bundle = quil_execution::message_envelope::CanonicalMessageBundle {
-                        requests: vec![Some(req)],
-                        timestamp,
-                    };
-                    match bundle.to_canonical_bytes() {
-                        Ok(bytes) => {
-                            let p2p = coverage_p2p.clone();
-                            tokio::spawn(async move {
-                                if let Err(e) = p2p
-                                    .publish(
-                                        quil_engine::bitmasks::GLOBAL_PROVER.to_vec(),
-                                        bytes,
-                                    )
-                                    .await
-                                {
-                                    warn!(error = %e, "coverage publish: GLOBAL_PROVER publish failed");
-                                }
-                            });
-                        }
-                        Err(e) => warn!(error = %e, "coverage publish: bundle encode failed"),
-                    }
-                });
-            // Per-worker state builder: each thread-mode worker opens
-            // its own RocksDB (resolved from db.worker_paths /
-            // worker_path_prefix / fallback) and builds its own
-            // clock_store, hypergraph CRDT, and execution engine on
-            // top. Master keeps its own global stores untouched.
-            let worker_db_base = config.db.path.clone();
-            let worker_paths_cfg = config.db.worker_paths.clone();
-            let worker_path_prefix_cfg = config.db.worker_path_prefix.clone();
-            let worker_state_builder: Arc<
-                dyn Fn(u32) -> std::result::Result<
-                    quil_engine::thread_worker::WorkerOwnedDeps,
-                    String,
-                > + Send
-                    + Sync,
-            > = Arc::new(move |core_id: u32| {
-                let path: std::path::PathBuf = {
-                    let idx = core_id.saturating_sub(1) as usize;
-                    if let Some(p) = worker_paths_cfg.get(idx).filter(|s| !s.is_empty()) {
-                        std::path::PathBuf::from(p)
-                    } else if !worker_path_prefix_cfg.is_empty() {
-                        std::path::PathBuf::from(
-                            worker_path_prefix_cfg.replace("%d", &core_id.to_string()),
-                        )
-                    } else {
-                        let base = if worker_db_base.is_empty() {
-                            std::path::PathBuf::from(".config/store")
-                        } else {
-                            std::path::PathBuf::from(&worker_db_base)
-                        };
-                        base.join(format!("worker-{}", core_id))
-                    }
-                };
-                std::fs::create_dir_all(&path).map_err(|e| {
-                    format!("worker {} mkdir {}: {e}", core_id, path.display())
-                })?;
-                let db = quil_store::RocksDb::open(&path).map_err(|e| {
-                    format!("worker {} open db {}: {e}", core_id, path.display())
-                })?;
-                let db_arc = Arc::new(db);
-                let clock_store: Arc<dyn quil_types::store::ClockStore> = Arc::new(
-                    quil_store::RocksClockStore::new(db_arc.inner()),
-                );
-                let hg_store: Arc<dyn quil_types::store::HypergraphStore> = Arc::new(
-                    quil_store::RocksHypergraphStore::new(db_arc.inner()),
-                );
-                let inclusion_prover: Arc<dyn quil_types::crypto::InclusionProver> =
-                    Arc::new(quil_crypto::KzgInclusionProver);
-                let crdt = Arc::new(quil_hypergraph::HypergraphCrdt::new(
-                    hg_store,
-                    inclusion_prover.clone(),
-                ));
-                // Workers don't sign or verify identities — a default
-                // key manager satisfies the execution engine's
-                // `KeyManager` requirement for state materialization.
-                let bls_constructor: Arc<dyn quil_types::crypto::BlsConstructor> =
-                    Arc::new(quil_crypto::Bls48581KeyConstructor);
-                let worker_key_manager: Arc<dyn quil_types::crypto::KeyManager> =
-                    Arc::new(quil_crypto::DefaultKeyManager::new(bls_constructor));
-                // Bulletproof prover is real; Decaf448 / circuit
-                // compiler still use the noop stubs (no production
-                // impl yet). See the analogous block in the master
-                // setup above for the rationale.
-                let bulletproof_prover: Arc<dyn quil_types::crypto::BulletproofProver> =
-                    Arc::new(quil_crypto::Decaf448BulletproofProver);
-                let decaf_constructor: Arc<dyn quil_types::crypto::DecafConstructor> =
-                    Arc::new(quil_execution::testing::NoopDecafConstructor);
-                let circuit_compiler: Arc<dyn quil_types::execution::CircuitCompiler> =
-                    Arc::new(quil_execution::testing::NoopCircuitCompiler);
-                let clock_store_for_exec: Arc<dyn quil_types::store::ClockStore> =
-                    clock_store.clone();
-                let hypergraph_resolver: Arc<dyn quil_execution::hypergraph_intrinsic::HypergraphConfigResolver> =
-                    Arc::new(quil_execution::testing::NoopHypergraphConfigResolver);
-                let exec_manager = Arc::new(
-                    quil_execution::ExecutionEngineManager::new(
-                        inclusion_prover.clone(),
-                        worker_key_manager,
-                        crdt.clone(),
-                        bulletproof_prover,
-                        decaf_constructor,
-                        circuit_compiler,
-                        clock_store_for_exec,
-                        hypergraph_resolver,
-                        true,
-                    ),
-                );
-                tracing::info!(
-                    core_id,
-                    path = %path.display(),
-                    "worker state initialized"
-                );
-                Ok(quil_engine::thread_worker::WorkerOwnedDeps {
-                    clock_store,
-                    hypergraph: crdt,
-                    execution_engine: exec_manager,
-                    inclusion_prover,
-                    // Each worker writes consensus + liveness state
-                    // into its own RocksDB. Mirrors the per-worker
-                    // clock/hypergraph stores above.
-                    kv_db: Some(db_arc.clone() as Arc<dyn quil_types::store::KvDb>),
-                })
-            });
-
-            thread_mgr.set_consensus_deps(quil_engine::thread_worker::WorkerConsensusDeps {
-                prover_registry: prover_registry.clone() as Arc<dyn quil_types::consensus::ProverRegistry>,
-                frame_prover: frame_prover.clone(),
-                message_collector: message_collector.clone(),
-                clock_store: clock_store.clone() as Arc<dyn quil_types::store::ClockStore>,
-                fee_manager: fee_manager.clone(),
-                local_prover_address: prover_address.to_vec(),
-                local_bls_pubkey: bls_pubkey.clone(),
-                bls_signer_factory: Arc::new(move || {
-                    fkm_for_factory.get_signer(quil_types::crypto::KeyType::Bls48581G1)
-                        .expect("BLS signer should be available")
-                }),
-                reward_greedy,
-                coverage_publish: Some(coverage_publish),
-                // Master's global state, used as fallback when the
-                // per-worker builder fails or isn't wired.
-                hypergraph: Some(crdt.clone()),
-                execution_engine: Some(exec_manager.clone()),
-                inclusion_prover: Some(inclusion_prover.clone()),
-                worker_init: Some(Arc::new(|core_id: u32| {
-                    crate::logging::set_worker_core_id(core_id);
-                    crate::logging::register_worker_log_file(core_id);
-                })),
-                worker_state_builder: Some(worker_state_builder),
-                // Master's RocksDB doubles as the persistent backing
-                // for app-shard `ConsensusState` / `LivenessState` —
-                // workers writing through the master path (no
-                // per-worker DB) land here. Per-worker builds can
-                // override via `WorkerOwnedDeps::kv_db`.
-                kv_db: Some(db_arc.clone() as Arc<dyn quil_types::store::KvDb>),
-            });
-            info!(
-                worker_cores = thread_mgr.num_worker_cores(),
-                "thread worker manager ready (local mode)"
-            );
-            // Drain `WorkerToMaster` events from in-process worker
-            // threads and forward to the master's BlossomSub publish
-            // path. `ShardFrameFinalized` becomes a
-            // `MessageBundle{Shard: header}` on `GLOBAL_PROVER`.
-            // Per-shard bitmask subscriptions are wired on
-            // `ShardActivated`; inbound routing dispatches by filter
-            // through `shard_engines` in the recv loop below.
-            if let Some(mut master_rx) = thread_mgr.take_master_rx() {
-                let drain_p2p = p2p_handle.clone();
-                let drain_cancel = token.clone();
-                let drain_shard_engines = shard_engines.clone();
-                let drain_halt = halt_state.clone();
-                tokio::spawn(async move {
-                    loop {
-                        tokio::select! {
-                            _ = drain_cancel.cancelled() => break,
-                            event = master_rx.recv() => {
-                                let Some(event) = event else { break; };
-                                use quil_engine::thread_worker::WorkerToMaster;
-                                // Each publish is dispatched as a fire-and-forget
-                                // task: the swarm's `publish().await` can block on
-                                // an internal mesh send, and back-pressure here
-                                // would propagate all the way to the per-shard
-                                // consensus event handler (engine→master event_tx
-                                // is bounded), stalling QC processing and
-                                // finalization.
-                                match event {
-                                    WorkerToMaster::ShardFrameFinalized {
-                                        core_id,
-                                        filter,
-                                        header_canonical_bytes,
-                                    } => {
-                                        // Drop reward-proof submissions during a coverage
-                                        // halt. The engine's per-message halt gates stop
-                                        // new consensus from advancing, but a finalize
-                                        // event already in-flight when the halt arrived
-                                        // can still race through and emit here. Suppress
-                                        // the publish so we don't credit shard work that
-                                        // shouldn't have happened during the halt window.
-                                        if drain_halt.any_halted() {
-                                            debug!(
-                                                core_id,
-                                                filter = %hex::encode(&filter),
-                                                "suppressing GLOBAL_PROVER publish — coverage halt active"
-                                            );
-                                            continue;
-                                        }
-                                        // Decode for a positive log line so the operator
-                                        // can see each rewardable proof going out. The
-                                        // bytes are consumed by `wrap` below; decode a
-                                        // borrowed view first.
-                                        if let Ok(h) =
-                                            quil_execution::global_intrinsic::frame_header::FrameHeader::from_canonical_bytes(
-                                                &header_canonical_bytes,
-                                            )
-                                        {
-                                            info!(
-                                                core_id,
-                                                filter = %hex::encode(&filter),
-                                                frame = h.frame_number,
-                                                rank = h.rank,
-                                                prover = %hex::encode(&h.prover),
-                                                "submitting reward proof to GLOBAL_PROVER"
-                                            );
-                                        }
-                                        let req = match quil_execution::message_envelope::CanonicalMessageRequest::wrap(
-                                            header_canonical_bytes,
-                                        ) {
-                                            Ok(r) => r,
-                                            Err(e) => {
-                                                warn!(core_id, filter = %hex::encode(&filter), error = %e,
-                                                    "shard finalize: bad FrameHeader bytes — dropping coverage publish");
-                                                continue;
-                                            }
-                                        };
-                                        let timestamp = std::time::SystemTime::now()
-                                            .duration_since(std::time::UNIX_EPOCH)
-                                            .unwrap_or_default()
-                                            .as_millis() as i64;
-                                        let bundle = quil_execution::message_envelope::CanonicalMessageBundle {
-                                            requests: vec![Some(req)],
-                                            timestamp,
-                                        };
-                                        match bundle.to_canonical_bytes() {
-                                            Ok(bytes) => {
-                                                let p2p = drain_p2p.clone();
-                                                let filter_owned = filter.clone();
-                                                tokio::spawn(async move {
-                                                    if let Err(e) = p2p
-                                                        .publish(
-                                                            quil_engine::bitmasks::GLOBAL_PROVER.to_vec(),
-                                                            bytes,
-                                                        )
-                                                        .await
-                                                    {
-                                                        warn!(core_id, filter = %hex::encode(&filter_owned),
-                                                            error = %e, "GLOBAL_PROVER publish failed");
-                                                    }
-                                                });
-                                            }
-                                            Err(e) => warn!(core_id, error = %e,
-                                                "shard finalize: bundle encode failed"),
-                                        }
-                                    }
-                                    WorkerToMaster::FrameProduced { core_id, filter, frame_data, .. } => {
-                                        // Per-shard frame bitmask = filter itself.
-                                        // Self-loopback is handled in thread_worker
-                                        // before we get here.
-                                        if drain_halt.any_halted() {
-                                            debug!(core_id, filter = %hex::encode(&filter),
-                                                "suppressing shard frame publish — coverage halt active");
-                                            continue;
-                                        }
-                                        let p2p = drain_p2p.clone();
-                                        tokio::spawn(async move {
-                                            if let Err(e) = p2p
-                                                .publish(
-                                                    quil_engine::bitmasks::shard_frame_bitmask(&filter),
-                                                    frame_data,
-                                                )
-                                                .await
-                                            {
-                                                warn!(core_id, filter = %hex::encode(&filter),
-                                                    error = %e, "shard frame publish failed");
-                                            }
-                                        });
-                                    }
-                                    WorkerToMaster::VoteProduced { core_id, filter, vote_data } => {
-                                        // Per-shard consensus bitmask = `0x00 || filter`.
-                                        if drain_halt.any_halted() {
-                                            debug!(core_id, filter = %hex::encode(&filter),
-                                                "suppressing shard vote publish — coverage halt active");
-                                            continue;
-                                        }
-                                        let p2p = drain_p2p.clone();
-                                        tokio::spawn(async move {
-                                            if let Err(e) = p2p
-                                                .publish(
-                                                    quil_engine::bitmasks::shard_consensus_bitmask(&filter),
-                                                    vote_data,
-                                                )
-                                                .await
-                                            {
-                                                warn!(core_id, filter = %hex::encode(&filter),
-                                                    error = %e, "shard vote publish failed");
-                                            }
-                                        });
-                                    }
-                                    WorkerToMaster::TimeoutProduced { core_id, filter, timeout_data } => {
-                                        if drain_halt.any_halted() {
-                                            debug!(core_id, filter = %hex::encode(&filter),
-                                                "suppressing shard timeout publish — coverage halt active");
-                                            continue;
-                                        }
-                                        let p2p = drain_p2p.clone();
-                                        tokio::spawn(async move {
-                                            if let Err(e) = p2p
-                                                .publish(
-                                                    quil_engine::bitmasks::shard_consensus_bitmask(&filter),
-                                                    timeout_data,
-                                                )
-                                                .await
-                                            {
-                                                warn!(core_id, filter = %hex::encode(&filter),
-                                                    error = %e, "shard timeout publish failed");
-                                            }
-                                        });
-                                    }
-                                    WorkerToMaster::ShardActivated { core_id, filter, handle } => {
-                                        // Push the current halt state to the
-                                        // freshly-activated engine before
-                                        // registering it. Without this the
-                                        // engine boots with halted=false and
-                                        // happily proposes frames during a
-                                        // network-wide halt window until the
-                                        // next halt-state transition arrives.
-                                        handle.set_halted(drain_halt.any_halted());
-                                        // Register the engine handle so the
-                                        // recv loop can dispatch peer
-                                        // messages to it.
-                                        {
-                                            let mut map = drain_shard_engines.write();
-                                            map.insert(filter.clone(), handle);
-                                        }
-                                        // Subscribe BlossomSub to the four
-                                        // per-shard bitmasks. Without these
-                                        // subscriptions our mesh peers won't
-                                        // forward shard traffic to us, so
-                                        // peer votes / proposals / frames /
-                                        // dispatches never reach the engine.
-                                        let p2p = drain_p2p.clone();
-                                        let filter_for_sub = filter.clone();
-                                        tokio::spawn(async move {
-                                            p2p.subscribe(quil_engine::bitmasks::shard_frame_bitmask(&filter_for_sub)).await;
-                                            p2p.subscribe(quil_engine::bitmasks::shard_consensus_bitmask(&filter_for_sub)).await;
-                                            p2p.subscribe(quil_engine::bitmasks::shard_prover_bitmask(&filter_for_sub)).await;
-                                            p2p.subscribe(quil_engine::bitmasks::shard_dispatch_bitmask(&filter_for_sub)).await;
-                                        });
-                                        info!(
-                                            core_id,
-                                            filter = %hex::encode(&filter),
-                                            "registered shard engine + subscribed per-shard bitmasks"
-                                        );
-                                    }
-                                    WorkerToMaster::ShardDeactivated { core_id, filter } => {
-                                        {
-                                            let mut map = drain_shard_engines.write();
-                                            map.remove(&filter);
-                                        }
-                                        let p2p = drain_p2p.clone();
-                                        let filter_for_sub = filter.clone();
-                                        tokio::spawn(async move {
-                                            p2p.unsubscribe(quil_engine::bitmasks::shard_frame_bitmask(&filter_for_sub)).await;
-                                            p2p.unsubscribe(quil_engine::bitmasks::shard_consensus_bitmask(&filter_for_sub)).await;
-                                            p2p.unsubscribe(quil_engine::bitmasks::shard_prover_bitmask(&filter_for_sub)).await;
-                                            p2p.unsubscribe(quil_engine::bitmasks::shard_dispatch_bitmask(&filter_for_sub)).await;
-                                        });
-                                        info!(
-                                            core_id,
-                                            filter = %hex::encode(&filter),
-                                            "deregistered shard engine + unsubscribed per-shard bitmasks"
-                                        );
-                                    }
-                                    WorkerToMaster::Ready { .. }
-                                    | WorkerToMaster::ShardHeartbeat { .. } => {
-                                        // No-op — informational only.
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    info!("worker→master drain task stopped");
-                });
-            }
-            // Restore persisted worker state (manually_managed flag +
-            // assigned filter) before any pre-allocation runs, so the
-            // operator's intent sticks across restarts.
-            //
-            // Archive mode skips the restore — `set_worker_filter`
-            // would otherwise spawn worker threads, and archives don't
-            // run app-shard workers. A subsequent return to non-archive
-            // will pick
-            // them up again because we don't delete them here.
-            let persisted = if archive_mode {
-                if !thread_mgr.load_all_persisted().is_empty() {
-                    info!("archive mode: skipping persisted worker restore");
-                }
-                Vec::new()
-            } else {
-                thread_mgr.load_all_persisted()
-            };
-            if !persisted.is_empty() {
-                info!(
-                    count = persisted.len(),
-                    "restoring persisted worker state from store"
-                );
-                for entry in persisted {
-                    // Resurrect the binding (filter pinned, no
-                    // consensus engine yet — `worker_allocator` will
-                    // re-attach the engine when the registry alloc
-                    // for this filter is observed Active).
-                    if !entry.filter.is_empty() {
-                        if let Err(e) = quil_engine::worker::WorkerManager::set_worker_filter(
-                            thread_mgr.as_ref(),
-                            entry.core_id,
-                            &entry.filter,
-                            false,
-                        ) {
-                            warn!(
-                                core_id = entry.core_id,
-                                error = %e,
-                                "failed to restore worker filter from store"
-                            );
-                        }
-                    }
-                    if entry.manually_managed {
-                        if let Err(e) = quil_engine::worker::WorkerManager::set_manually_managed(
-                            thread_mgr.as_ref(),
-                            entry.core_id,
-                            true,
-                        ) {
-                            warn!(
-                                core_id = entry.core_id,
-                                error = %e,
-                                "failed to restore manually_managed flag"
-                            );
-                        }
-                    }
-                    if entry.pending_filter_frame > 0 {
-                        let _ = quil_engine::worker::WorkerManager::set_pending_filter_frame(
-                            thread_mgr.as_ref(),
-                            entry.core_id,
-                            entry.pending_filter_frame,
-                        );
-                    }
-                }
-            }
-            thread_mgr as Arc<dyn quil_engine::worker::WorkerManager>
-        };
-
-    // Pre-allocate idle workers for each available core so they're
-    // online from startup. Workers start idle (empty filter) and get
-    // assigned shards by the lifecycle when join proposals are accepted.
-    //
-    // Archive mode skips this entirely. Per the architecture
-    // (re-stated at the `frame_materializer` block below): archives
-    // materialize global frames; workers materialize app-shard frames
-    // — a separate role. An archive node spawning app-shard workers
-    // would be every-role-at-once, which is wrong: an archive's job
-    // is to retain global history and serve sync, not to compete
-    // for shard rewards. The other gates (lifecycle.evaluate,
-    // worker_allocator.on_new_frame) are also archive-skipped in
-    // their respective call sites below.
-    if !archive_mode {
-        let num_cores = match worker_manager.check_workers_connected() {
-            Ok(ids) => ids.len() as u32,
-            Err(_) => 0,
-        };
-        // If no workers exist yet, create them for cores 1..N
-        if num_cores == 0 {
-            let total = std::thread::available_parallelism()
-                .map(|n| n.get() as u32)
-                .unwrap_or(4);
-            let worker_count = total.saturating_sub(1).max(1); // reserve core 0 for master
-            for core_id in 1..=worker_count {
-                if let Err(e) = worker_manager.allocate_worker(core_id, &[]) {
-                    warn!(core_id, error = %e, "failed to pre-allocate idle worker");
-                }
-            }
-            info!(workers = worker_count, "pre-allocated idle workers");
-        }
-    } else {
-        info!("archive mode: skipping worker pre-allocation (archives don't run app-shard workers)");
-    }
-
-    // Apply `engine.data_worker_filters` from YAML config. Runs AFTER
-    // persisted-restore and idle pre-allocation:
-    //   * fresh node pins config filters with manually_managed=true;
-    //   * restart with prior persisted/gRPC assignment keeps it
-    //     (persisted wins).
-    // Skipped in archive mode for the same reason as pre-allocation.
-    if !archive_mode {
-        let cfg_filters = &config.engine.data_worker_filters;
-        let stats = quil_engine::worker_allocator::apply_config_worker_filters(
-            worker_manager.as_ref(),
-            cfg_filters,
-        );
-        if !cfg_filters.is_empty() {
-            info!(
-                declared = cfg_filters.len(),
-                applied = stats.applied,
-                skipped_existing = stats.skipped_existing,
-                skipped_missing_core = stats.skipped_missing_core,
-                skipped_empty = stats.skipped_empty,
-                invalid = stats.invalid,
-                "applied engine.data_worker_filters"
-            );
-        }
-    } else if !config.engine.data_worker_filters.is_empty() {
-        info!(
-            declared = config.engine.data_worker_filters.len(),
-            "archive mode: ignoring engine.data_worker_filters (archives don't run app-shard workers)"
-        );
-    }
-
-    // Publish the worker_manager handle to the PeerInfo broadcaster.
-    // From this point on, every PeerInfo tick advertises a
-    // per-worker reachability for each running worker with a
-    // non-empty filter. Thread-mode workers (the default) share the
-    // master's addresses; process-mode workers (when
-    // `engine.data_worker_p2p_multiaddrs` or
-    // `engine.data_worker_stream_multiaddrs` is configured) advertise
-    // their own ports. See
-    // `quil_p2p::peer_info::build_worker_reachability` for the
-    // selection rules.
-    let _ = pi_worker_manager.set(worker_manager.clone());
-
-    // Worker allocator — reconciles registry vs running workers
-    let worker_allocator = Arc::new(quil_engine::worker_allocator::WorkerAllocator::new(
-        worker_manager.clone(),
-        prover_registry.clone() as Arc<dyn quil_types::consensus::ProverRegistry>,
-        prover_address.to_vec(),
-    ));
-
-    // Compute the config-derived seniority estimate from the mainnet
-    // compat table. Uses our local libp2p peer ID plus any peer IDs
-    // derived from the configs listed in
-    // `engine.multisig_prover_enrollment_paths`. Result is cached on
-    // the allocator; lifecycle consults it when deciding whether a
-    // seniority merge would raise our on-chain seniority. Computed
-    // only on mainnet (P2P.Network == 0).
-    if config.p2p.network == 0 {
-        let mut peer_ids: Vec<String> = Vec::new();
-        let pk_bytes = hex::decode(&config.p2p.peer_priv_key).unwrap_or_default();
-        if pk_bytes.len() >= 57 {
-            let mut seed = [0u8; 57];
-            seed.copy_from_slice(&pk_bytes[..57]);
-            let pubkey = quil_p2p::ed448_identity::derive_public_key(&seed);
-            let peer_id = quil_p2p::ed448_identity::peer_id_from_ed448_pubkey(&pubkey);
-            peer_ids.push(bs58::encode(&peer_id).into_string());
-        }
-        for extra_path in &config.engine.multisig_prover_enrollment_paths {
-            let path = std::path::PathBuf::from(extra_path);
-            match quil_config::load_config(&path) {
-                Ok(extra_cfg) => {
-                    if let Ok(bytes) = hex::decode(&extra_cfg.p2p.peer_priv_key) {
-                        if bytes.len() >= 57 {
-                            let mut seed = [0u8; 57];
-                            seed.copy_from_slice(&bytes[..57]);
-                            let pubkey = quil_p2p::ed448_identity::derive_public_key(&seed);
-                            let peer_id = quil_p2p::ed448_identity::peer_id_from_ed448_pubkey(&pubkey);
-                            peer_ids.push(bs58::encode(&peer_id).into_string());
-                        }
-                    }
-                }
-                Err(e) => warn!(
-                    path = %extra_path,
-                    error = %e,
-                    "could not load multisig prover enrollment config"
-                ),
-            }
-        }
-        let estimate = quil_execution::seniority_compat::get_aggregated_seniority(&peer_ids);
-        worker_allocator.set_config_seniority_estimate(estimate);
-        info!(
-            local_peer_ids = peer_ids.len(),
-            aggregated_seniority = estimate,
-            "computed config-derived seniority estimate"
-        );
-    }
-
-    // Proactive worker allocation reconcile — fire as soon as the
-    // background prover registry refresh has data, independent of
-    // any archive sync. This is the path that pulls workers out of
-    // "filter pinned, consensus deferred" state on startup based on
-    // the LOCALLY persisted prover registry, rather than waiting
-    // for the first global frame to arrive from the archive poller
-    // (which can take 5+ minutes during PeerInfo discovery).
-    //
-    // Archive nodes don't host app-shard workers, so skip the call
-    // there — matches the gate inside the archive poller.
-    if !archive_mode {
-        let wa_for_init = worker_allocator.clone();
-        let pr_for_init = prover_registry.clone();
-        let lhf_for_init = last_global_head_frame.clone();
-        let token_for_init = token.clone();
-        tokio::spawn(async move {
-            // The background refresh at startup is spawn_blocking and
-            // typically finishes in well under a second. Poll up to
-            // ~5s for `distinct_provers() > 0`; on a fresh node with
-            // an empty registry this just gives up silently and the
-            // archive-poller path picks it up later.
-            for _ in 0..50 {
-                if token_for_init.is_cancelled() { return; }
-                let count = pr_for_init.read(|r| r.distinct_provers());
-                if count > 0 { break; }
-                tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-            }
-            let frame = lhf_for_init.load(std::sync::atomic::Ordering::Relaxed);
-            if frame > 0 {
-                if let Err(e) = wa_for_init.on_new_frame(frame) {
-                    warn!(
-                        error = %e,
-                        frame,
-                        "early worker reconcile failed"
-                    );
-                } else {
-                    info!(frame, "early worker reconcile complete (off local registry)");
-                }
-            }
-        });
-    }
-
-    // Shared slot for the consensus event-loop handle, populated by the
-    // sync task once a genesis frame is in the store. The receive loop
-    // and lifecycle pipeline read from it to feed inbound proposals/QCs/TCs
-    // back into the HotStuff event loop.
-    let consensus_handle: Arc<std::sync::OnceLock<
-        quil_engine::consensus_types::GlobalEventLoopHandle,
-    >> = Arc::new(std::sync::OnceLock::new());
-
-    // Per-rank vote aggregator. Populated alongside the handle by
-    // `activate_consensus`. The receive loop feeds inbound
-    // ProposalVote + GlobalProposal messages in so votes accumulate
-    // toward a quorum certificate, which is then submitted back to
-    // the event loop via the shared handle.
-    let vote_aggregator: Arc<std::sync::OnceLock<
-        Arc<quil_engine::vote_aggregation::VoteAggregation>,
-    >> = Arc::new(std::sync::OnceLock::new());
-
-    // Per-rank timeout aggregator. Same lifecycle as the vote aggregator
-    // but for TimeoutState messages — produces TCs (and partial TCs)
-    // from aggregated timeout signatures.
-    let timeout_aggregator: Arc<std::sync::OnceLock<
-        Arc<quil_engine::timeout_aggregation::TimeoutAggregation>,
-    >> = Arc::new(std::sync::OnceLock::new());
-
-    // Prover lifecycle coordinator — evaluates join/confirm/leave on each frame.
-    // Pulls cooldown state off the WorkerAllocator (single source of truth).
-    let strategy = if reward_greedy {
-        quil_engine::provers::proposer::Strategy::RewardGreedy
-    } else {
-        quil_engine::provers::proposer::Strategy::DataGreedy
-    };
-    let lifecycle_inner = quil_engine::provers::lifecycle::ProverLifecycle::new(
-        prover_address.to_vec(),
-        worker_allocator.clone(),
-        halt_state.clone(),
-        current_frame.clone(),
-        strategy,
-    );
-    // Testnet/devnet bootstraps drop the join-confirm window from
-    // mainnet's 360 frames (one hour at 10s/frame) to a handful of
-    // frames so a local smoke test sees a join → confirm cycle in
-    // a couple of minutes. Mainnet (network = 0) keeps the full
-    // 360-frame protocol value.
-    // CLI `--network` is the source of truth — the YAML's `p2p.network`
-    // is left at 0 in our test configs because the same files are used
-    // for mainnet runs. Use the run-time-supplied `network` arg so a
-    // single config file can be reused across networks.
-    if network != 0 {
-        const TESTNET_CONFIRM_WINDOW_FRAMES: u64 = 10;
-        lifecycle_inner.set_confirm_window_frames(TESTNET_CONFIRM_WINDOW_FRAMES);
-        // The lifecycle setting controls *when the local node submits*
-        // a Confirm. The materializer's `validate_confirm_timing`
-        // independently enforces that the recipient ledger has waited
-        // long enough — its default (360..720) is mainnet-correct. For
-        // testnet we have to override that window too, otherwise every
-        // submitted Confirm is rejected as "must wait 360 frames after
-        // join" until 360 frames have elapsed (an hour) — exactly the
-        // wait the lifecycle override is meant to avoid.
-        quil_execution::global_intrinsic::verify::set_confirm_window_frames(
-            TESTNET_CONFIRM_WINDOW_FRAMES,
-            // Use a generous upper bound so a slow follower can still
-            // confirm before the window expires.
-            TESTNET_CONFIRM_WINDOW_FRAMES * 72, // 720 ÷ 360 × 10 = 20 → 720
-        );
-        info!(
-            network,
-            confirm_window_frames = TESTNET_CONFIRM_WINDOW_FRAMES,
-            "testnet/devnet: using shortened prover confirm window",
-        );
-    }
-    let prover_lifecycle = Arc::new(lifecycle_inner);
-    // Wire the shards store so `evaluate` can discover shards that
-    // have no allocations yet — calls `RangeAppShards` on the local
-    // store.
-    prover_lifecycle.set_shards_store(
-        shards_store.clone() as Arc<dyn quil_types::store::ShardsStore>,
-    );
-
-    // Per the architecture: archives materialize global frames;
-    // workers materialize app-shard frames (a separate path);
-    // non-archive masters only consume the materialized state via
-    // sync from archives and do not materialize themselves.
-    //
-    // Archive nodes get the FrameMaterializer (the canonical
-    // post-finalize processor: commits the hypergraph, verifies the
-    // prover root, processes message bundles, prunes orphan joins,
-    // evicts inactive provers, persists alt-shard updates, publishes
-    // post-materialize snapshots) plus the full frame-header deps on
-    // the intrinsic so `invoke_frame_header` mutates state on
-    // shard-coverage ingest (LastActiveFrameNumber advance + reward
-    // distribution). That full install already wires `frame_prover`.
-    //
-    // Non-archive masters install only `frame_prover` — the minimum
-    // their archive-poller path needs so `process_global_frame` →
-    // `validate_message` → the intrinsic's `TYPE_PROVER_JOIN` arm
-    // doesn't fail closed with "frame_prover not installed".
-    let reward_issuer: Arc<dyn quil_types::consensus::RewardIssuance> =
-        Arc::new(quil_engine::OptRewardIssuance);
-    if archive_mode {
-        let bls_for_intrinsic: Arc<dyn quil_types::crypto::BlsConstructor> =
-            Arc::new(quil_crypto::Bls48581KeyConstructor);
-        exec_manager.install_global_frame_header_deps(
-            prover_registry.clone() as Arc<dyn quil_types::consensus::ProverRegistry>,
-            reward_issuer.clone(),
-            bls_for_intrinsic,
-            inclusion_prover.clone(),
-            frame_prover.clone(),
-        ).expect("failed to install global frame header dependencies");
-    } else {
-        exec_manager.install_global_frame_prover(frame_prover.clone()).expect("failed to install global frame prover");
-    }
-    let frame_materializer: Option<Arc<quil_engine::frame_materializer::FrameMaterializer>> =
-        if archive_mode {
-            let m = quil_engine::frame_materializer::FrameMaterializer::new(
-                exec_manager.clone(),
-                prover_registry.clone() as Arc<dyn quil_types::consensus::ProverRegistry>,
-                clock_store.clone() as Arc<dyn quil_types::store::ClockStore>,
-                crdt.clone(),
-                hg_store.clone() as Arc<dyn quil_types::store::HypergraphStore>,
-                reward_issuer,
-                prover_address.to_vec(),
-                archive_mode,
-            )
-            .with_eviction_registry(prover_registry.clone())
-            .with_rocks_hg_store(hg_store.clone())
-            .with_current_frame(current_frame.clone());
-            Some(Arc::new(m))
-        } else {
-            None
-        };
 
     // ---------------------------------------------------------------
     // 6. Message receive loop
